+++ NeuralNetworkLib/Synapse.cs
@@ -2,15 +2,15 @@
 {
     internal class Synapse
     {
-        public int Weight { get; private set; }
+        public double Weight { get; private set; }
 
-        public int? Value { get;  set; }
+        public double? Value { get;  set; }
 
         public Neuron Source { get;}
 
         public Neuron Destination { get;}
 
-        public Synapse(Neuron source, Neuron destination, int weight)
+        public Synapse(Neuron source, Neuron destination, double weight)
         {
             this.Weight = weight;
             this.Value = null;
